Adopt SqliteEventSpool, delete DiskQueue/EventSpool
MixpanelClient now constructs SqliteEventSpool; the DiskQueue-backed
EventSpool and its dependency are removed entirely. Contract tests
collapse to the single remaining engine (28 tests, was 56 across two
engines); the DiskQueue-vs-SQLite responsiveness characterization test
is removed along with the engine it characterized.

Verified never-throws by inspection rather than porting DiskQueue's
disk-IO-failure tests: all six public IEventSpool members on
SqliteEventSpool (Enqueue, ProcessBatchAsync, TrimExpired, Purge,
ApproximateCount, ApproximateBytes) wrap their body in try/catch and
log-and-swallow -- the property IEventSpool documents ("never throws")
holds structurally, not by test coverage of a specific fault path.

dotnet list package --include-transitive confirms DiskQueue is gone from
all three TFMs. 102/102 Mixpanel tests green on net462 and net8.0; core
(19+14) unaffected.
