+++ src/DesktopAnalyticsTests.Mixpanel/TrackResponsivenessTests.cs
@@ -9,13 +9,12 @@ using NUnit.Framework;
 
 namespace DesktopAnalyticsTests
 {
-	// EventSpool (DiskQueue) holds its global _sync semaphore across the network send, so a
-	// UI-thread Track() call blocks for the WHOLE send -- up to ~45s against a stalled connection
-	// in production. SqliteEventSpool claims a batch under a short lock, releases it, sends, then
-	// resolves the claim under a second short lock, so Track() stays responsive throughout.
-	// NB: the stall is EventSpool's lock, not a DiskQueue limitation -- DiskQueue supports
-	// concurrent sessions (probed 2026-07-16), so dropping that lock would also fix it. See the
-	// class remarks on SqliteEventSpool.
+	// SqliteEventSpool claims a batch under a short lock, releases it, sends with nothing held,
+	// then resolves the claim under a second short lock, so Track() stays responsive even while a
+	// send is genuinely in flight -- never blocking for the whole network round trip. See the
+	// class remarks on SqliteEventSpool for the full rationale (including the now-removed
+	// DiskQueue-backed EventSpool's contrasting behavior, which this used to characterize
+	// alongside the fix).
 	[TestFixture]
 	public class TrackResponsivenessTests
 	{
@@ -116,63 +115,6 @@ namespace DesktopAnalyticsTests
 				Assert.AreEqual(0, spool.ApproximateCount,
 					"the Track() call made while the send was in flight must have been spooled, " +
 					"not dropped, and must drain cleanly afterward");
-				Assert.AreEqual(2, client.Statistics.Submitted);
-			}
-		}
-
-		// ---- THE DEFECT (characterization, not a requirement): EventSpool blocks ---------------
-		//
-		// This test PASSES against today's EventSpool behavior on purpose -- it pins the documented
-		// defect (see MixpanelClientTests.ConcurrentTrack_WhileDrainInProgress_..., whose comments
-		// describe Track() threads "piling up" behind the lock) so it is visible in the test suite
-		// rather than hidden in a comment. Delete it if EventSpool is dropped -- or if EventSpool's
-		// global lock is narrowed so it stops blocking, in which case this test should start failing
-		// and be deleted rather than "fixed".
-		[Test]
-		public async Task Track_WhileSendInFlight_BlocksUntilSendCompletes_DiskQueue()
-		{
-			using (var spool = new EventSpool(_spoolDir, 1000))
-			{
-				var sender = new BlockingUntilSignaledSender();
-				var client = new MixpanelClient();
-				client.InitializeForTest(spool, sender, batchSize: 5);
-
-				client.Track("user-1", "Seed", null);
-
-				var drainTask = client.DrainOnceAsync(); // Blocks inside the sender until Release().
-				await WaitForSendToStart(sender);
-
-				// Run Track() on its own thread so this test can observe it NOT finishing while the
-				// send is blocked, rather than deadlocking the test itself.
-				var trackFinished = new ManualResetEventSlim(false);
-				var trackThread = new Thread(() =>
-				{
-					client.Track("user-1", "WhileSendInFlight", null);
-					trackFinished.Set();
-				});
-				trackThread.Start();
-
-				// EventSpool's transactional session holds its lock across the whole send, so
-				// Track() must still be blocked after a short window -- this is the defect, pinned.
-				var finishedWhileBlocked = trackFinished.Wait(TimeSpan.FromMilliseconds(300));
-				Assert.IsFalse(finishedWhileBlocked,
-					"characterizes today's known defect: EventSpool holds its lock across the " +
-					"network send, so Track() blocks for the whole send instead of returning " +
-					"promptly. If this ever fails, EventSpool has stopped blocking Track() and this " +
-					"test (and the bug it documents) should be deleted, not fixed.");
-
-				// Release the blocked send so the Track() thread (and the drain) can finally
-				// complete -- otherwise this test would hang forever.
-				sender.Release();
-
-				Assert.IsTrue(trackFinished.Wait(TimeSpan.FromSeconds(5)),
-					"Track() must complete once the send is released");
-				trackThread.Join();
-				await drainTask;
-
-				Assert.AreEqual(1, client.Statistics.Succeeded);
-				await client.DrainOnceAsync();
-				Assert.AreEqual(0, spool.ApproximateCount);
 				Assert.AreEqual(2, client.Statistics.Submitted);
 			}
 		}
